backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
gunicorn
Pillow
djangorestframework-simplejwt